show downvotes to admins; move voting into module
