+++ pages/answers/[answerId].tsx
@@ -212,6 +212,14 @@ const SingleAnswer = () => {
                       <span className="material-icons">delete</span>
                     </button>
                     <span className="ml-6">IP: ({answer.ip})</span>
+                    {answer.vote === -1 && (
+                      <button
+                        className="ml-4 text-red-600"
+                        title="Downvote"
+                      >
+                        <span className="material-icons">thumb_down</span>
+                      </button>
+                    )}
                   </>
                 )}
               </div>
